# PERF, BENCH: Fix performance issue when indexing into non-unique DatetimeIndex/PeriodIndex.