@pep8speaks 2019-06-30
Hello @qwhelan! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:





* In the file [`pandas/core/indexes/period.py`](https://github.com/pandas-dev/pandas/blob/a68caa2ab04d80074c38fc18d97a3e5ca039dea5/pandas/core/indexes/period.py):

> [Line 604:80](https://github.com/pandas-dev/pandas/blob/a68caa2ab04d80074c38fc18d97a3e5ca039dea5/pandas/core/indexes/period.py#L604): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (82 > 79 characters)